refactor: Rename table-engine to mito
