+++ Cargo.toml
@@ -32,7 +32,7 @@ members = [
     "src/storage",
     "src/store-api",
     "src/table",
-    "src/table-engine",
+    "src/mito",
 ]
 
 [profile.release]
