+++ src/mito/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "table-engine"
+name = "mito"
 version = "0.1.0"
 edition = "2021"
 license = "Apache-2.0"
