+++ src/mito/README.md
@@ -1,0 +1,3 @@
+# Mito
+
+Mito is GreptimeDB's default table engine.
