+++ src/sql/Cargo.toml
@@ -15,4 +15,4 @@ itertools = "0.10"
 once_cell = "1.10"
 snafu = { version = "0.7", features = ["backtraces"] }
 sqlparser = "0.15.0"
-table-engine = { path = "../table-engine" }
+mito = { path = "../mito" }
